Fixed CAI API plurality issue for networkInterfaces
The CAI API object for compute instances has the block `networkInterfaces`
and within that `accessConfigs`, but the validator for vm_external_ip as well as
multiple fixtures use the singular version. Another issue with the vm_external_ip
validator was that at the time of the terraform plan, an external_ip will not
necessarily exist, so checking for the existence of an `access_config` is a more
stable solution.